requirements.txt
streamlit
pymongo
python-dotenv
groq
pandas
PyPDF2
pdfplumber
pytesseract
Pillow
nltk
bcrypt
language_tool_python
